code example from 1/17
we got swerve working off this except it would only go in one direction and would never stop, The encoder was reading the right amount but would stop reading if it got within the range of -420 - 420. You could change speed though.  I also added a desired angle variable to the smart dashboard